move prune cmd tests to own module
