+++ setup.cfg
@@ -159,6 +159,7 @@ per_file_ignores =
     src/borg/testsuite/archive.py:E128,W504
     src/borg/testsuite/archiver/__init__.py:E128,E501,E722,F401,F405,F811
     src/borg/testsuite/archiver/debug_cmds.py:E501
+    src/borg/testsuite/archiver/prune_cmd.py:F405
     src/borg/testsuite/benchmark.py:F401,F811
     src/borg/testsuite/chunker.py:E501,F405
     src/borg/testsuite/chunker_pytest.py:F401
